Fix LevelDevel
the problem is actually somewhere else, where current position in mem stream is used to calculate remaining amount of points to load. Since we never rewinded mem, it was positioned at the end of the memory buffer and amount calculated to 0.

Before the introduction of that code position didn't matter because actual reading of the points seeks to the right place and all points were read. The new code has to do with reading slices that are smaller than the memory buffer.